+++ src/main/kotlin/picklab/backend/auth/application/OAuthUserInfoMapper.kt
@@ -3,6 +3,10 @@ package picklab.backend.auth.application
 import com.fasterxml.jackson.databind.JsonNode
 import org.springframework.stereotype.Component
 import picklab.backend.auth.domain.OAuthUserInfo
+import picklab.backend.auth.infrastructure.GithubUserInfo
+import picklab.backend.auth.infrastructure.GoogleUserInfo
+import picklab.backend.auth.infrastructure.KakaoUserInfo
+import picklab.backend.auth.infrastructure.NaverUserInfo
 
 @Component
 class OAuthUserInfoMapper {
