+++ src/main/kotlin/picklab/backend/auth/domain/AuthToken.kt
@@ -1,6 +1,6 @@
 package picklab.backend.auth.domain
 
-data class TokenResponse(
+data class AuthToken(
     val accessToken: String,
     val refreshToken: String,
 )
